add track-record value translation tag (#5305)
* add track-record value translation tag

* add context to the track record tag labels

* Update bg_selector_tags.py
