+++ network-api/networkapi/buyersguide/templates/fragments/it_uses.html
@@ -1,4 +1,4 @@
-{% load yes_no i18n %}
+{% load bg_selector_tags i18n %}
 
 <div class="it-uses d-flex flex-column">
   <div class="title">
