+++ network-api/networkapi/buyersguide/templates/fragments/signup_requirement.html
@@ -1,4 +1,4 @@
-{% load yes_no i18n %}
+{% load bg_selector_tags i18n %}
 
 <div class="signup-requirement d-flex flex-column">
   <div class="title">
